Rename SDAccel header to OpenCL to dodge future renaming by Xilinx, and update tests accordingly. Remove separate SDAccel CMake file.
